fix: use int64 instead of int for numeric values from queries

+++ src/metrics/metrics.go
@@ -118,7 +118,7 @@ func populateWaitTimeMetrics(instanceEntity *integration.Entity, connection *con
 
 		metrics := []struct {
 			metricName  string
-			metricValue int
+			metricValue int64
 			metricType  metric.SourceType
 		}{
 			{
